pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aerich>=0.9.2",
    "asyncclick>=8.3.0.7",
    "asyncpg>=0.31.0",
    "anyio>=4.4.0",
    "black>=25.11.0",
    "coverage>=7.12.0",
    "dictdiffer>=0.9.0",
    "dotenv>=0.9.9",
    "fastapi>=0.122.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "ciso8601>=2.3.0",
    "pydantic-settings>=2.3.4",
    "pypika-tortoise>=0.6.3",
    "pytz>=2025.2",
    "starlette>=0.37.0",
    "tortoise-orm>=0.25.1",
    "typing-extensions>=4.12.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]

[tool.aerich]
tortoise_orm = "app.main.TORTOISE_ORM"
location = "./migrations"
